+++ src/main/scala/scalismo/faces/landmarks/LandmarkDetectionMap.scala
@@ -15,76 +15,6 @@
  */
 package scalismo.faces.landmarks
 
-import scalismo.faces.image.{PixelImage, PixelImageDomain}
-import scalismo.faces.image.filter.GeneralMaxConvolution
-import scalismo.faces.sampling.face.evaluators.PointEvaluators.IsotropicGaussianPointEvaluator
-import scalismo.geometry.{Point1D, _1D}
+import scalismo.faces.image.PixelImage
 
-
-/**
-  * LandmarkDetectionMap is a tagged PixelImage[Double].
-  */
-trait LandmarkDetectionMap extends ((Int, Int) => Double) {
-  def tag: String
-  def domain: PixelImageDomain
-}
-
-/**
-  * The implementation of the LandmarkDetectionMap.
-  */
-private class LandmarkDetectionMapImplementation(
-  override val tag: String,
-  private val detectionMap: PixelImage[Double]
-) extends LandmarkDetectionMap {
-  override def domain: PixelImageDomain = detectionMap.domain
-  def apply(x: Int, y: Int): Double = detectionMap(x, y)
-}
-
-/**
-  * The factory object for a LandmarkDetectionMap.
-  */
-object LandmarkDetectionMap {
-
-  /**
-    * Constructor where everything needs to be specified.
-    */
-  def apply(tag: String, map: PixelImage[Double], stddevNoiseModel: Double, fp: Double, fn: Double): LandmarkDetectionMap = {
-    val correctedMap = correctFalsePositiveAndFalseNegativeRates(map, fp, fn)
-    val precalculatedMapIncludingNoiseModel = precalculateIsotropicGaussianNoise(correctedMap, stddevNoiseModel)
-    new LandmarkDetectionMapImplementation(tag, precalculatedMapIncludingNoiseModel)
-  }
-
-  /**
-    * Constructor if the correction for the false-positive and false-negative rates needs to be ommited.
-    */
-  def fromCorrectedDetection(tag: String, correctedMap: PixelImage[Double], stddevNoiseModel: Double): LandmarkDetectionMap = {
-    val precalculatedMapIncludingNoiseModel = precalculateIsotropicGaussianNoise(correctedMap, stddevNoiseModel)
-    new LandmarkDetectionMapImplementation(tag, precalculatedMapIncludingNoiseModel)
-  }
-
-  /**
-    * Construction for a LandmarkDetectionMap without correction nor a precalculated noise model.
-    * @note We do not recommend to use this constructor unless you know what you are doing.
-    */
-  def fromCorrectedPrecalculatedMap(tag: String, precalculatedMapIncludingNoiseModel: PixelImage[Double]): LandmarkDetectionMap = {
-    new LandmarkDetectionMapImplementation(tag, precalculatedMapIncludingNoiseModel)
-  }
-
-  /**
-    * Correctes for the false-positive and false-negative rates of the detector in the detection map.
-    */
-  private def correctFalsePositiveAndFalseNegativeRates(in: PixelImage[Double], falsePositiveRate: Double, falseNegativeRate: Double): PixelImage[Double] = {
-    in.map(e => e * (1.0 - (falsePositiveRate + falseNegativeRate)) + falseNegativeRate)
-  }
-
-  /**
-    * Precalculates a new map with included isotropic Gaussian noise model.
-    */
-  private def precalculateIsotropicGaussianNoise(detectionMap: PixelImage[Double], stddevNoiseModel: Double) = {
-    GeneralMaxConvolution.separable2D(
-      detectionMap,
-      IsotropicGaussianPointEvaluator[_1D](stddevNoiseModel).toDistributionEvaluator(Point1D(0f))
-    )
-  }
-
-}
+case class LandmarkDetectionMap(tag: String, values: PixelImage[Double])
